Renamed abstractions package to base
