+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/activities/SingleFragmentActivity.java
@@ -9,7 +9,7 @@ import android.support.v4.app.NavUtils;
 import android.view.MenuItem;
 
 import com.ninetwozero.bf4intel.R;
-import com.ninetwozero.bf4intel.abstractions.BaseIntelActivity;
+import com.ninetwozero.bf4intel.base.BaseIntelActivity;
 import com.ninetwozero.bf4intel.factories.FragmentFactory;
 
 public class SingleFragmentActivity extends BaseIntelActivity {
