+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/fragments/BattleFeedFragment.java
@@ -5,7 +5,7 @@ import android.os.Bundle;
 import android.view.*;
 
 import com.ninetwozero.bf4intel.R;
-import com.ninetwozero.bf4intel.abstractions.BaseListFragment;
+import com.ninetwozero.bf4intel.base.BaseListFragment;
 import com.ninetwozero.bf4intel.activities.SingleFragmentActivity;
 import com.ninetwozero.bf4intel.adapters.FeedAdapter;
 import com.ninetwozero.bf4intel.factories.FragmentFactory;
