+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/fragments/BattleFeedPostingFragment.java
@@ -7,7 +7,7 @@ import android.view.*;
 import android.widget.EditText;
 
 import com.ninetwozero.bf4intel.R;
-import com.ninetwozero.bf4intel.abstractions.BaseFragment;
+import com.ninetwozero.bf4intel.base.BaseFragment;
 import com.ninetwozero.bf4intel.activities.SingleFragmentActivity;
 
 public class BattleFeedPostingFragment extends BaseFragment {
