+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/fragments/NewsArticleFragment.java
@@ -10,7 +10,7 @@ import android.widget.EditText;
 import android.widget.ListView;
 
 import com.ninetwozero.bf4intel.R;
-import com.ninetwozero.bf4intel.abstractions.BaseListFragment;
+import com.ninetwozero.bf4intel.base.BaseListFragment;
 import com.ninetwozero.bf4intel.adapters.NewsItemAdapter;
 
 import java.util.ArrayList;
