+++ BF4Intel/src/main/java/com/ninetwozero/bf4intel/fragments/ThreadCreationFragment.java
@@ -11,7 +11,7 @@ import android.view.ViewGroup;
 import android.widget.EditText;
 
 import com.ninetwozero.bf4intel.R;
-import com.ninetwozero.bf4intel.abstractions.BaseFragment;
+import com.ninetwozero.bf4intel.base.BaseFragment;
 import com.ninetwozero.bf4intel.activities.SingleFragmentActivity;
 
 public class ThreadCreationFragment extends BaseFragment {
